gc: consolidate remaining repr collectability coverage
